+++ tests/integration_tests/adapters/betfair/test_kit.py
@@ -52,6 +52,7 @@ from nautilus_trader.live.data_engine import LiveDataEngine
 from nautilus_trader.model.commands.trading import CancelOrder
 from nautilus_trader.model.commands.trading import ModifyOrder
 from nautilus_trader.model.commands.trading import SubmitOrder
+from nautilus_trader.model.data.tick import TradeTick
 from nautilus_trader.model.enums import OrderSide
 from nautilus_trader.model.enums import TimeInForce
 from nautilus_trader.model.events.order import OrderAccepted
@@ -64,6 +65,7 @@ from nautilus_trader.model.identifiers import VenueOrderId
 from nautilus_trader.model.instruments.betting import BettingInstrument
 from nautilus_trader.model.objects import Price
 from nautilus_trader.model.objects import Quantity
+from nautilus_trader.model.orderbook.data import OrderBookData
 from nautilus_trader.model.orders.limit import LimitOrder
 from nautilus_trader.model.orders.market import MarketOrder
 from nautilus_trader.persistence.config import PersistenceConfig
@@ -536,11 +538,11 @@ class BetfairTestStubs:
             venues=[BetfairTestStubs.betfair_venue_config()],
             data=[
                 base_data_config.replace(
-                    data_cls_path="nautilus_trader.model.data.tick.TradeTick",
+                    data_cls=TradeTick,
                     instrument_id=instrument_id,
                 ),
                 base_data_config.replace(
-                    data_cls_path="nautilus_trader.model.orderbook.data.OrderBookData",
+                    data_cls=OrderBookData,
                     instrument_id=instrument_id,
                 ),
             ],