+++ tests/unit_tests/common/test_common_actor.py
@@ -112,12 +112,9 @@ class TestActor:
         self.data_engine.start()
         self.exec_engine.start()
 
-    def test_actor_fullname(self):
-        # Arrange
-        actor = Actor(config=ActorConfig(component_id=self.component_id))
-
-        # Act, Assert
-        assert actor.fullname() == "nautilus_trader.common.actor.Actor"
+    def test_actor_fully_qualified_name(self):
+        # Arrange, Act, Assert
+        assert Actor.fully_qualified_name() == "nautilus_trader.common.actor.Actor"
 
     def test_id(self):
         # Arrange, Act